Added Thumbnail generator
Given an image (/assets/test.png), a thumbnail is generated when you visit /assets/test.<size>.png whereas the limiting <size> is:
- a width: 300w
- or a height: 300h
This works with PNG, GIF and JPEG files.
Once generated, the thumbnail is saved and will not be generated again.
The allowed <size> values can be set in include/config.inc.php.